# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Keramik Store Platform API
  description: |
    Backend REST API for ceramic tile e-commerce platform with RAG-powered product assistant and Midtrans payment integration.
    
    ## Features
    - **E-commerce Core**: Product catalog, inventory, cart, orders, and payments
    - **RAG Assistant**: Document-based product knowledge Q&A using vector search and LLM
    - **Payment Integration**: Midtrans Snap payment with webhook synchronization
    - **Discount System**: Product discounts, vouchers, and promotional codes
    
    ## Authentication
    The API supports multiple authentication methods with comprehensive security features:
    
    ### Email/Password Authentication
    - **Registration**: Create account via `/v1/auth/register` with email verification
    - **Email Verification**: Verify email with token sent to inbox
    - **Login**: Authenticate via `/v1/auth/login` (requires verified email)
    - **Password Management**: 
      - Forgot password flow with email reset link
      - Reset password with token
      - Change password for authenticated users
    
    ### Social Authentication (OAuth2)
    - **Google OAuth2**: Social login via `/v1/auth/google`
    - **Facebook OAuth2**: Social login via `/v1/auth/facebook`
    
    ### Token Management
    - **JWT Tokens**: Access tokens (short-lived) and refresh tokens (long-lived)
    - **Token Refresh**: Renew access token via `/v1/auth/refresh`
...
    
...
    
    ### OAuth2 Flow
...
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      description: |
        Create a new customer account. After successful registration, a verification email will be sent to the provided email address.
        User must verify their email before they can login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 255
                  description: Must contain at least 8 characters
                  example: SecurePass123!
                name:
                  type: string
                  maxLength: 255
                  example: John Doe
      responses:
        '201':
          description: User registered successfully. Verification email sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registration successful. Please check your email to verify your account.
                  userId:
                    type: string
                    format: uuid
                  email:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify email address
      description: Verify user's email address using the token sent via email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Email verification token from the email
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email verified successfully. You can now login.
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/resend-verification:
    post:
      tags: [Auth]
      summary: Resend verification email
      description: Resend email verification link to user's email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verification email sent. Please check your inbox.
        '400':
          description: Email already verified or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests. Please wait before requesting again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login user
      description: |
        Authenticate user and return JWT tokens.
        User must have verified their email address before login.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidCredentials:
                  value:
                    statusCode: 401
                    message: Invalid email or password
                    error: Unauthorized
                emailNotVerified:
                  value:
                    statusCode: 401
                    message: Please verify your email before logging in
                    error: Unauthorized
  /v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Send password reset email with reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Password reset email sent (always returns success even if email doesn't exist for security)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: If an account with that email exists, a password reset link has been sent.
        '429':
          description: Too many requests. Please wait before requesting again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: Reset user password using the token from forgot-password email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  description: Password reset token from email
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                newPassword:
                  type: string
                  format: password
                  minLength: 8
                  description: New password (minimum 8 characters)
                  example: NewSecurePass123!
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset successful. You can now login with your new password.
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                email:
                  type: string
                  format: email
                  maxLength: 255
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/users:
    get:
      tags: [Users]
      summary: List all users [ADMIN]
      description: Get paginated list of users (admin only)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
        - name: search
          in: query
          description: Search by name or email
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID [ADMIN]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: User retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== PRODUCTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /v1/auth/login
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, STAFF, CUSTOMER]
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Token expiration in seconds
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        provider:
          type: string
          nullable: true
          enum: [local, google, facebook]
          description: Authentication provider (local for email/password, google, or facebook)
        providerId:
          type: string
          nullable: true
          description: OAuth provider user ID (Google ID or Facebook ID)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedUsers:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/User'

    # ==================== PRODUCTS ====================
...
    PaginationMeta:
      type: object
      properties:
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasNextPage:
              type: boolean
            hasPreviousPage:
              type: boolean
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 400
            message: ["price must be a positive number"]
            error: Bad Request
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 401
            message: Unauthorized
            error: Unauthorized
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 403
            message: Forbidden resource
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 404
            message: Resource not found
            error: Not Found